Chore/user

+++ src/user/user.service.ts
@@ -24,24 +24,31 @@ export class UserService {
           id: userId,
         },
       });
-      delete user.createdAt;
+
+      if (user) {
+            delete user.createdAt;
       delete user.password;
-
-      const preferences = await this.prisma.preferences.findUnique({
-        where: { userId },
-      });
-      delete preferences.id;
-      delete preferences.userId;
-
-      const creditCardDetails = await this.prisma.creditCardDetails.findUnique({
-        where: { userId },
-      });
-      delete creditCardDetails.id;
-      delete creditCardDetails.userId;
-      const payload = {
-        user: { ...user, preferences, creditCardDetails },
-      };
-      return this.resHandler.requestSuccessful({ res, payload });
+      
+        const preferences = await this.prisma.preferences.findUnique({
+          where: { userId },
+        });
+        delete preferences.id;
+        delete preferences.userId;
+   
+        const creditCardDetails =
+          await this.prisma.creditCardDetails.findUnique({
+            where: { userId },
+          });
+        delete creditCardDetails.id;
+        delete creditCardDetails.userId;
+        const payload = {
+          user: { ...user, preferences, creditCardDetails },
+        };
+        return this.resHandler.requestSuccessful({ res, payload });
+      } 
+      else {
+        return this.resHandler.clientError(res, 'This user does not exist!');
+      }
     } catch (err) {
       return this.resHandler.serverError(res, 'Error getting user details');
     }
@@ -56,14 +63,19 @@ export class UserService {
    */
   async updateUserBio(userId: string, dto: UpdateBioDto, res: Response) {
     try {
-      await this.prisma.user.update({
-        where: { id: userId },
-        data: { ...dto },
-      });
-      return this.resHandler.requestSuccessful({
-        res,
-        message: 'User details updated successfully',
-      });
+      const user = await this.prisma.user.findUnique({ where: { id: userId } });
+      if (user) {
+        await this.prisma.user.update({
+          where: { id: userId },
+          data: { ...dto },
+        });
+        return this.resHandler.requestSuccessful({
+          res,
+          message: 'User details updated successfully',
+        });
+      } else {
+        return this.resHandler.clientError(res, 'This user does not exist!');
+      }
     } catch (err) {
       return this.resHandler.serverError(res, 'Error updating user details');
     }
@@ -81,14 +93,25 @@ export class UserService {
     res: Response,
   ) {
     try {
-      await this.prisma.preferences.update({
+      const preferences = await this.prisma.preferences.findUnique({
         where: { userId },
-        data: { ...dto },
       });
-      return this.resHandler.requestSuccessful({
-        res,
-        message: 'Preferences updated successfully',
-      });
+
+      if (preferences) {
+        await this.prisma.preferences.update({
+          where: { userId },
+          data: { ...dto },
+        });
+        return this.resHandler.requestSuccessful({
+          res,
+          message: 'Preferences updated successfully',
+        });
+      } else {
+        return this.resHandler.clientError(
+          res,
+          'There are no preferences associated with this user',
+        );
+      }
     } catch (err) {
       console.log(err);
       return this.resHandler.serverError(
@@ -110,14 +133,25 @@ export class UserService {
     res: Response,
   ) {
     try {
-      await this.prisma.creditCardDetails.update({
+      const creditCard = this.prisma.creditCardDetails.findUnique({
         where: { userId },
-        data: { ...dto },
       });
-      return this.resHandler.requestSuccessful({
-        res,
-        message: 'Credit card details updated successfully',
-      });
+
+      if (creditCard) {
+        await this.prisma.creditCardDetails.update({
+          where: { userId },
+          data: { ...dto },
+        });
+        return this.resHandler.requestSuccessful({
+          res,
+          message: 'Credit card details updated successfully',
+        });
+      } else {
+        return this.resHandler.clientError(
+          res,
+          'There are no credit card details associated with this uesr',
+        );
+      }
     } catch (err) {
       return this.resHandler.serverError(
         res,
@@ -133,17 +167,22 @@ export class UserService {
    */
   async deleteUser(userId: string, res: Response) {
     try {
-      // first delete preferences
-      await this.prisma.preferences.delete({ where: { userId } });
-      // then delete credit card
-      await this.prisma.creditCardDetails.delete({ where: { userId } });
-      // then delete user
-      await this.prisma.user.delete({ where: { id: userId } });
+      const user = await this.prisma.user.findUnique({ where: { id: userId } });
+      if (user) {
+        // first delete preferences
+        await this.prisma.preferences.delete({ where: { userId } });
+        // then delete credit card
+        await this.prisma.creditCardDetails.delete({ where: { userId } });
+        // then delete user
+        await this.prisma.user.delete({ where: { id: userId } });
 
-      return this.resHandler.requestSuccessful({
-        res,
-        message: 'User deleted successfully',
-      });
+        return this.resHandler.requestSuccessful({
+          res,
+          message: 'User deleted successfully',
+        });
+      } else {
+        return this.resHandler.clientError(res, 'User does not exist!');
+      }
     } catch (err) {
       return this.resHandler.serverError(res, 'Error deleting user');
     }
